Move plugin helpers from init, move constants from funcs to top level
